+++ config/resolve_test.go
@@ -37,6 +37,7 @@ func captureLog(t *testing.T, logger **stdlog.Logger) *logtest.Buffer {
 	resetLegacyRootAgentsWarnings()
 	resetConfigAliasWarnings()
 	resetShellValueWarnings()
+	resetInRepoUnknownLeafWarnings()
 	var buf logtest.Buffer
 	old := *logger
 	*logger = stdlog.New(&buf, "", 0)